+++ libtpms/src/tpm2/Object.c
@@ -159,7 +159,7 @@ BOOL ObjectIsSequence(OBJECT* object  // IN: handle to be checked
 {
     pAssert(object != NULL);
     return (object->attributes.hmacSeq == SET || object->attributes.hashSeq == SET
-	    || object->attributes.eventSeq == SET);
+	    || object->attributes.eventSeq == SET || object->attributes.pqcSeq == SET);
 }
 
 //*** HandleToObject()